Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
redis==5.2.1
httpx==0.28.1
python-multipart==0.0.20
